+++ docs/document/content/user-manual/shardingsphere-jdbc/configuration/java-api/replica.cn.md
@@ -5,6 +5,6 @@ weight = 5
 
 ## 配置入口
 
-类名称：org.apache.shardingsphere.replica.api.config.ReplicaRuleConfiguration
+类名称：org.apache.shardingsphere.replication.consensus.api.config.ReplicaRuleConfiguration
 
 TODO
