--- shardingsphere-features/shardingsphere-consensus-replication/shardingsphere-consensus-replication-common/src/main/java/org/apache/shardingsphere/replica/yaml/swapper/ReplicaActualTableRuleConfigurationYamlSwapper.java
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.replica.yaml.swapper;
-
-import org.apache.shardingsphere.infra.yaml.swapper.YamlSwapper;
-import org.apache.shardingsphere.replica.api.config.ReplicaActualTableRuleConfiguration;
-import org.apache.shardingsphere.replica.yaml.config.YamlReplicaActualTableRuleConfiguration;
-
-/**
- * Replica actual table rule configuration YAML swapper.
- */
-public final class ReplicaActualTableRuleConfigurationYamlSwapper implements YamlSwapper<YamlReplicaActualTableRuleConfiguration, ReplicaActualTableRuleConfiguration> {
-    
-    @Override
-    public YamlReplicaActualTableRuleConfiguration swapToYamlConfiguration(final ReplicaActualTableRuleConfiguration data) {
-        YamlReplicaActualTableRuleConfiguration result = new YamlReplicaActualTableRuleConfiguration();
-        result.setPhysicsTable(data.getPhysicsTable());
-        result.setReplicaGroupId(data.getReplicaGroupId());
-        result.setReplicaPeers(data.getReplicaPeers());
-        result.setDataSourceName(data.getDataSourceName());
-        return result;
-    }
-    
-    @Override
-    public ReplicaActualTableRuleConfiguration swapToObject(final YamlReplicaActualTableRuleConfiguration yamlConfiguration) {
-        return new ReplicaActualTableRuleConfiguration(yamlConfiguration.getPhysicsTable(), yamlConfiguration.getReplicaGroupId(),
-                yamlConfiguration.getReplicaPeers(), yamlConfiguration.getDataSourceName());
-    }
-}
