+++ shardingsphere-features/shardingsphere-consensus-replication/shardingsphere-consensus-replication-common/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.yaml.swapper.YamlRuleConfigurationSwapper
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.replica.yaml.swapper.ReplicaRuleConfigurationYamlSwapper
+org.apache.shardingsphere.replication.consensus.yaml.swapper.ReplicaRuleConfigurationYamlSwapper
