+++ shardingsphere-features/shardingsphere-consensus-replication/shardingsphere-consensus-replication-executor/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.executor.sql.raw.execute.callback.RawExecutorCallback
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.replica.execute.callback.DefaultReplicaExecutorCallback
+org.apache.shardingsphere.replication.consensus.execute.callback.DefaultReplicaExecutorCallback
