--- shardingsphere-features/shardingsphere-consensus-replication/shardingsphere-consensus-replication-route/src/main/java/org/apache/shardingsphere/replica/route/engine/ReplicaGroup.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.replica.route.engine;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.apache.shardingsphere.infra.route.context.RawGroup;
-
-/**
- * Replica group.
- */
-@AllArgsConstructor
-@Getter
-public final class ReplicaGroup implements RawGroup {
-    
-    public static final String BLANK_REPLICA_GROUP_KEY = "Nil";
-    
-    private final String physicsTable;
-    
-    private final String replicaGroupId;
-    
-    private final String replicaPeers;
-    
-    private final String dataSourceName;
-}
