+++ shardingsphere-features/shardingsphere-consensus-replication/shardingsphere-consensus-replication-route/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.route.decorator.RouteDecorator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.replica.route.engine.ReplicaRouteDecorator
+org.apache.shardingsphere.replication.consensus.route.engine.ReplicaRouteDecorator
